+++ src/RcsbCollectTools/SequenceCollector/SequenceCollector.ts
@@ -18,6 +18,7 @@ import {
 } from "./SequenceCollectorInterface";
 import {Subject} from "rxjs";
 import {ObservableHelper} from "../../RcsbUtils/ObservableHelper";
+import {ExternalTrackBuilderInterface} from "../FeatureTools/ExternalTrackBuilderInterface";
 
 export interface AlignedObservedRegion extends AlignedRegion {
     unModelled?:boolean;
@@ -43,6 +44,7 @@ export class SequenceCollector implements SequenceCollectorInterface{
     private readonly targetsSubject: Subject<Array<string>> = new Subject<Array<string>>();
     private alignmentResponse: AlignmentResponse;
     private readonly alignmentResponseSubject: Subject<AlignmentResponse> = new Subject<AlignmentResponse>();
+    private externalTrackBuilder: ExternalTrackBuilderInterface;
 
     private tagObservedRegions: (region: AlignedRegion, commonContext: TranslateContextInterface) => Array<AlignedObservedRegion> = (region: AlignedRegion, commonContext: TranslateContextInterface) => {
         return [{...region,unModelled:false}];
@@ -116,6 +118,10 @@ export class SequenceCollector implements SequenceCollectorInterface{
     public setPolymerEntityInstanceTranslator(p: PolymerEntityInstanceTranslate): void {
         this.helper.setPolymerEntityInstanceTranslator(p);
         this.polymerEntityInstanceTranslator = p;
+    }
+
+    public setExternalTrackBuilder(externalTrackBuilder: ExternalTrackBuilderInterface): void {
+        this.externalTrackBuilder = externalTrackBuilder;
     }
 
     private complete(){