+++ src/RcsbFvWeb/RcsbFvModule/RcsbFvEntity.ts
@@ -1,20 +1,18 @@
 import {SequenceReference, Source} from "@rcsb/rcsb-api-tools/build/RcsbGraphQL/Types/Borrego/GqlTypes";
 import {RcsbFvAbstractModule} from "./RcsbFvAbstractModule";
 import {RcsbFvModuleBuildInterface} from "./RcsbFvModuleInterface";
-import {RcsbFvRowConfigInterface} from "@rcsb/rcsb-saguaro";
-import {SequenceCollectorDataInterface} from "../../RcsbCollectTools/SequenceCollector/SequenceCollector";
 
 export class RcsbFvEntity extends RcsbFvAbstractModule {
 
-    public async build(buildConfig: RcsbFvModuleBuildInterface): Promise<void> {
+    protected async protectedBuild(buildConfig: RcsbFvModuleBuildInterface): Promise<void> {
         const entityId: string = buildConfig.entityId;
         const source: Array<Source> = buildConfig.additionalConfig?.filters? buildConfig.additionalConfig.sources : [Source.PdbEntity, Source.Uniprot];
-        const seqResult:SequenceCollectorDataInterface = await this.sequenceCollector.collect({
+        this.alignmentTracks = await this.sequenceCollector.collect({
             queryId: entityId,
             from: SequenceReference.PdbEntity,
             to: SequenceReference.Uniprot
         });
-        const annResult: Array<RcsbFvRowConfigInterface> = await this.annotationCollector.collect({
+        this.annotationTracks = await this.annotationCollector.collect({
                 queryId: entityId,
                 reference: SequenceReference.PdbEntity,
                 sources:source,
@@ -23,9 +21,14 @@ export class RcsbFvEntity extends RcsbFvAbstractModule {
         });
         this.boardConfigData.length = this.sequenceCollector.getSequenceLength();
         this.boardConfigData.includeAxis = true;
-        this.rowConfigData = !buildConfig.additionalConfig?.hideAlignments ? seqResult.sequence.concat(seqResult.alignment).concat(annResult) : seqResult.sequence.concat(seqResult.alignment).concat(annResult);
-        await this.display();
         return void 0;
+    }
+
+    protected concatAlignmentAndAnnotationTracks(buildConfig: RcsbFvModuleBuildInterface): void {
+        this.rowConfigData = !buildConfig.additionalConfig?.hideAlignments ?
+            this.alignmentTracks.sequence.concat(this.alignmentTracks.alignment).concat(this.annotationTracks)
+            :
+            this.alignmentTracks.sequence.concat(this.alignmentTracks.alignment).concat(this.annotationTracks);
     }
 
 }