+++ src/RcsbFvWeb/RcsbFvModule/RcsbFvInstance.ts
@@ -1,29 +1,35 @@
 import {SequenceReference, Source} from "@rcsb/rcsb-api-tools/build/RcsbGraphQL/Types/Borrego/GqlTypes";
 import {RcsbFvAbstractModule} from "./RcsbFvAbstractModule";
-import {RcsbFvModuleBuildInterface, RcsbFvModuleInterface} from "./RcsbFvModuleInterface";
-import {RcsbFv, RcsbFvRowConfigInterface} from "@rcsb/rcsb-saguaro";
-import {SequenceCollectorDataInterface} from "../../RcsbCollectTools/SequenceCollector/SequenceCollector";
+import {RcsbFvModuleBuildInterface} from "./RcsbFvModuleInterface";
 
 export class RcsbFvInstance extends RcsbFvAbstractModule {
 
-    public async build(buildConfig: RcsbFvModuleBuildInterface): Promise<void> {
+    protected async protectedBuild(buildConfig: RcsbFvModuleBuildInterface): Promise<void> {
         const instanceId: string = buildConfig.instanceId;
         const source: Array<Source> = [Source.PdbEntity, Source.PdbInstance, Source.Uniprot];
-        const seqResult:SequenceCollectorDataInterface = await this.sequenceCollector.collect({
+
+        this.alignmentTracks = await this.sequenceCollector.collect({
             queryId: instanceId,
             from: SequenceReference.PdbInstance,
             to: SequenceReference.Uniprot});
 
-        const annResult: Array<RcsbFvRowConfigInterface> = await this.annotationCollector.collect({
+        this.annotationTracks = await this.annotationCollector.collect({
                 queryId: instanceId,
                 reference: SequenceReference.PdbInstance,
-                sources:source});
+                sources:source
+        });
 
         this.boardConfigData.length = this.sequenceCollector.getSequenceLength();
         this.boardConfigData.includeAxis = true;
-        this.rowConfigData = !buildConfig.additionalConfig?.hideAlignments ? seqResult.sequence.concat(seqResult.alignment).concat(annResult) : seqResult.sequence.concat(annResult);
-        await this.display();
         return void 0;
+    }
+
+    protected concatAlignmentAndAnnotationTracks(buildConfig:RcsbFvModuleBuildInterface): void {
+        this.rowConfigData =
+            !buildConfig.additionalConfig?.hideAlignments ?
+                this.alignmentTracks.sequence.concat(this.alignmentTracks.alignment).concat(this.annotationTracks)
+                :
+                this.alignmentTracks.sequence.concat(this.annotationTracks);
     }
 
 }